moved proof height ante handler from common to seperate package
